djangoAPI/ibmLMS/requirements.txt
asgiref==3.2.7
astroid==2.4.2
Babel==2.8.0
bson==0.5.8
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.2.13
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-rest-swagger==2.2.0
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djongo==1.3.1
djoser==2.0.3
docutils==0.15.2
ibm-cos-sdk==2.6.3
ibm-cos-sdk-core==2.6.3
ibm-cos-sdk-s3transfer==2.6.3
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
phonenumbers==8.12.5
psycopg2-binary==2.8.5
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
